添加项目成员管理权限判断

+++ lion-console/src/main/java/com/dianping/lion/web/action/system/ProductAction.java
@@ -27,10 +27,12 @@ import com.dianping.lion.entity.Product;
 import com.dianping.lion.entity.Project;
 import com.dianping.lion.entity.Team;
 import com.dianping.lion.entity.User;
+import com.dianping.lion.exception.NoPrivilegeException;
 import com.dianping.lion.service.ProductService;
 import com.dianping.lion.service.ProjectService;
 import com.dianping.lion.service.TeamService;
 import com.dianping.lion.service.UserService;
+import com.dianping.lion.util.SecurityUtils;
 import com.dianping.lion.web.action.common.AbstractLionAction;
 
 @SuppressWarnings("serial")
@@ -66,6 +68,9 @@ public class ProductAction extends AbstractLionAction implements ServletRequestA
 	}
 	
 	public String productAddSubmit(){
+		if (!hasEditPrivilege()) {
+			throw NoPrivilegeException.INSTANCE;
+		}
 		Product product = new Product();
 		product.setName(name);
 		product.setProductLeaderId(productLeaderId);
@@ -88,6 +93,9 @@ public class ProductAction extends AbstractLionAction implements ServletRequestA
 	}
 	
 	public String productEditSubmit(){
+		if (!hasEditPrivilege()) {
+			throw NoPrivilegeException.INSTANCE;
+		}
 		Product product = new Product();
 		product.setId(id);
 		product.setName(name);
@@ -101,16 +109,23 @@ public class ProductAction extends AbstractLionAction implements ServletRequestA
 	}
 	
 	public String deleteProductAjax() {
-		List<Project>  productProjects =projectService.getProjectsByProduct(id);
+		if (!hasEditPrivilege()) {
+			throw NoPrivilegeException.INSTANCE;
+		}
+		List<Project>  productProjects = projectService.getProjectsByProduct(id);
 		if(productProjects.size() > 0) {
 			Product product = productService.findProductByID(id);
-			errorMessage = product.getName()+errorMessage;
+			errorMessage = product.getName() + errorMessage;
 			return ERROR;
 		} else {
 			productService.delete(id);
 			reInitiate();
 			return SUCCESS;
 		}
+	}
+	
+	public boolean hasEditPrivilege() {
+		return this.privilegeDecider.hasEditProjectPrivilege(SecurityUtils.getCurrentUserId());
 	}
 	
 	public String productList(){
